+++ packages/block-editor/src/components/media-placeholder/index.js
@@ -18,7 +18,7 @@ import MediaUploadCheck from '../media-upload/check';
 import URLPopover from '../url-popover';
 import { store as blockEditorStore } from '../../store';
 import { parseDropEvent } from '../use-on-block-drop';
-import { getComputedAcceptAttribute } from './utils';
+import { getComputedAcceptAttribute, isMediaDropEligible } from './utils';
 
 const noop = () => {};
 
@@ -372,20 +372,13 @@ export function MediaPlaceholder( {
 			<DropZone
 				onFilesDrop={ onFilesUpload }
 				onDrop={ handleBlocksDrop }
-				isEligible={ ( dataTransfer ) => {
-					const prefix = 'wp-block:core/';
-					const types = [];
-					for ( const type of dataTransfer.types ) {
-						if ( type.startsWith( prefix ) ) {
-							types.push( type.slice( prefix.length ) );
-						}
-					}
-					return (
-						types.every( ( type ) =>
-							allowedTypes.includes( type )
-						) && ( multiple ? true : types.length === 1 )
-					);
-				} }
+				isEligible={ ( dataTransfer ) =>
+					isMediaDropEligible(
+						dataTransfer.types,
+						allowedTypes,
+						multiple
+					)
+				}
 			/>
 		);
 	};